refactor: organise the files better
